COM-1695: change feature organization
